+++ packages/services/pkg/grpc/relay.go
@@ -299,7 +299,7 @@ func (server *ecsRelayServer) VerifySufficientBalance(client *relay.Client, addr
 		server.logger.Info("fetched up-to-date balance for account", zap.String("address", address), zap.Uint64("balance", balance))
 
 		// Update the "cached" balance on the client, which helps us know whether to keep checking or not.
-		client.SetHasSufficientBalance(balance > uint64(server.config.MinAccountBalance))
+		client.SetHasSufficientBalance(balance > server.config.MinAccountBalance)
 
 		if balance == 0 {
 			return fmt.Errorf("client with address %s has insufficient balance to push messages via relay", address)
